+++ breeze-data/src/main/kotlin/com/windea/breezeframework/data/generators/code/SqlGenerator.kt
@@ -15,7 +15,7 @@ object SqlGenerator : TextGenerator {
 	 * 输入文本的格式：`#/{database}/{table}/[]/{columns}/{column}`。
 	 */
 	fun generateSqlData(inputText: String, inputType: DataType = DataType.Yaml): String {
-		val rawInputMap = inputType.loader.fromString(inputText)
+		val rawInputMap = inputType.serializer.fromString(inputText)
 		val inputMap = rawInputMap as SqlDataMap
 		val databaseName = inputMap.keys.first()
 		val database = inputMap.values.first()
@@ -30,7 +30,7 @@ object SqlGenerator : TextGenerator {
 			insert into $tableName ($columnNamesSnippet) values
 			${table.joinToString(",\n", "", ";\n") { data ->
 				val columnsSnippet = data.values.joinToString {
-					it.toString().wrapQuote("'").unescape() ?: "null"
+					it.toString().wrapQuote("'").unescape()
 				}
 				
 				"""  ($columnsSnippet)"""