+++ breeze-data/src/main/kotlin/com/windea/breezeframework/data/serializers/json/FastJsonSerializer.kt
@@ -1,53 +1,34 @@
 package com.windea.breezeframework.data.serializers.json
 
+import com.alibaba.fastjson.*
 import java.io.*
 
-class FastJsonSerializer : JsonSerializer {
-	override fun <T : Any> load(string: String, type: Class<T>): T {
+class FastJsonSerializer : JsonSerializer<FastJsonSerializer, Any> {
+	override fun configure(handler: (Any) -> Unit): FastJsonSerializer {
 		TODO("not implemented")
+	}
+	
+	override fun <T : Any> load(string: String, type: Class<T>): T {
+		return JSON.parseObject(string, type)
 	}
 	
 	override fun <T : Any> load(file: File, type: Class<T>): T {
-		TODO("not implemented")
+		return file.readText().let { JSON.parseObject(it, type) }
 	}
 	
 	override fun <T : Any> load(reader: Reader, type: Class<T>): T {
-		TODO("not implemented")
-	}
-	
-	override fun loadList(string: String): List<Any?> {
-		TODO("not implemented")
-	}
-	
-	override fun loadList(file: File): List<Any?> {
-		TODO("not implemented")
-	}
-	
-	override fun loadList(reader: Reader): List<Any?> {
-		TODO("not implemented")
-	}
-	
-	override fun loadMap(file: File): Map<String, Any?> {
-		TODO("not implemented")
-	}
-	
-	override fun loadMap(string: String): Map<String, Any?> {
-		TODO("not implemented")
-	}
-	
-	override fun loadMap(reader: Reader): Map<String, Any?> {
-		TODO("not implemented")
+		return reader.readText().let { JSON.parseObject(it, type) }
 	}
 	
 	override fun <T : Any> dump(data: T): String {
-		TODO("not implemented")
+		return JSON.toJSONString(data, true)
 	}
 	
 	override fun <T : Any> dump(data: T, file: File) {
-		TODO("not implemented")
+		return JSON.writeJSONString(file.writer(), data)
 	}
 	
 	override fun <T : Any> dump(data: T, writer: Writer) {
-		TODO("not implemented")
+		return JSON.writeJSONString(writer, data)
 	}
 }