+++ breeze-data/src/main/kotlin/com/windea/breezeframework/data/serializers/xml/JacksonXmlSerializer.kt
@@ -1,54 +1,38 @@
 package com.windea.breezeframework.data.serializers.xml
 
+import com.fasterxml.jackson.dataformat.xml.*
 import java.io.*
 
-class JacksonXmlSerializer : XmlSerializer {
+class JacksonXmlSerializer : XmlSerializer<JacksonXmlSerializer, XmlMapper> {
+	private val mapper = XmlMapper()
+	
+	
+	/**配置持久化选项。这个方法必须首先被调用。*/
+	override fun configure(handler: (XmlMapper) -> Unit): JacksonXmlSerializer {
+		return this.also { handler(mapper) }
+	}
+	
 	override fun <T : Any> load(string: String, type: Class<T>): T {
-		TODO("not implemented")
+		return mapper.readValue(string, type)
 	}
 	
 	override fun <T : Any> load(file: File, type: Class<T>): T {
-		TODO("not implemented")
+		return mapper.readValue(file, type)
 	}
 	
 	override fun <T : Any> load(reader: Reader, type: Class<T>): T {
-		TODO("not implemented")
-	}
-	
-	override fun loadList(string: String): List<Any?> {
-		TODO("not implemented")
-	}
-	
-	override fun loadList(file: File): List<Any?> {
-		TODO("not implemented")
-	}
-	
-	override fun loadList(reader: Reader): List<Any?> {
-		TODO("not implemented")
-	}
-	
-	override fun loadMap(file: File): Map<String, Any?> {
-		TODO("not implemented")
-	}
-	
-	override fun loadMap(string: String): Map<String, Any?> {
-		TODO("not implemented")
-	}
-	
-	override fun loadMap(reader: Reader): Map<String, Any?> {
-		TODO("not implemented")
+		return mapper.readValue(reader, type)
 	}
 	
 	override fun <T : Any> dump(data: T): String {
-		TODO("not implemented")
+		return mapper.writeValueAsString(data)
 	}
 	
 	override fun <T : Any> dump(data: T, file: File) {
-		TODO("not implemented")
+		return mapper.writeValue(file, data)
 	}
 	
 	override fun <T : Any> dump(data: T, writer: Writer) {
-		TODO("not implemented")
+		return mapper.writeValue(writer, data)
 	}
-	
 }