+++ breeze-data/src/main/kotlin/com/windea/breezeframework/data/serializers/yaml/JacksonYamlSerializer.kt
@@ -1,78 +1,62 @@
 package com.windea.breezeframework.data.serializers.yaml
 
+import com.fasterxml.jackson.dataformat.yaml.*
 import java.io.*
 
-class JacksonYamlSerializer : YamlSerializer {
+class JacksonYamlSerializer : YamlSerializer<JacksonYamlSerializer, YAMLMapper> {
+	private val mapper = YAMLMapper()
+	
+	
+	/**配置持久化选项。这个方法必须首先被调用。*/
+	override fun configure(handler: (YAMLMapper) -> Unit): JacksonYamlSerializer {
+		return this.also { handler(mapper) }
+	}
+	
 	override fun <T : Any> load(string: String, type: Class<T>): T {
-		TODO("not implemented")
+		return mapper.readValue(string, type)
 	}
 	
 	override fun <T : Any> load(file: File, type: Class<T>): T {
-		TODO("not implemented")
+		return mapper.readValue(file, type)
 	}
 	
 	override fun <T : Any> load(reader: Reader, type: Class<T>): T {
-		TODO("not implemented")
+		return mapper.readValue(reader, type)
 	}
 	
-	override fun loadList(string: String): List<Any?> {
-		TODO("not implemented")
+	override fun loadAll(string: String): List<Any> {
+		throw UnsupportedOperationException("Could not find suitable methods to delegate in yamlMapper.")
 	}
 	
-	override fun loadList(file: File): List<Any?> {
-		TODO("not implemented")
+	override fun loadAll(file: File): List<Any> {
+		throw UnsupportedOperationException("Could not find suitable methods to delegate in yamlMapper.")
 	}
 	
-	override fun loadList(reader: Reader): List<Any?> {
-		TODO("not implemented")
+	override fun loadAll(reader: Reader): List<Any> {
+		throw UnsupportedOperationException("Could not find suitable methods to delegate in yamlMapper.")
 	}
-	
-	override fun loadMap(file: File): Map<String, Any?> {
-		TODO("not implemented")
-	}
-	
-	override fun loadMap(string: String): Map<String, Any?> {
-		TODO("not implemented")
-	}
-	
-	override fun loadMap(reader: Reader): Map<String, Any?> {
-		TODO("not implemented")
-	}
-	
-	override fun <T : Any> loadAll(string: String, type: Class<T>): List<T> {
-		TODO("not implemented")
-	}
-	
-	override fun <T : Any> loadAll(file: File, type: Class<T>): List<T> {
-		TODO("not implemented")
-	}
-	
-	override fun <T : Any> loadAll(reader: Reader, type: Class<T>): List<T> {
-		TODO("not implemented")
-	}
-	
 	
 	override fun <T : Any> dump(data: T): String {
-		TODO("not implemented")
+		return mapper.writeValueAsString(data)
 	}
 	
 	override fun <T : Any> dump(data: T, file: File) {
-		TODO("not implemented")
+		return mapper.writeValue(file, data)
 	}
 	
 	override fun <T : Any> dump(data: T, writer: Writer) {
-		TODO("not implemented")
+		return mapper.writeValue(writer, data)
 	}
 	
 	override fun <T : Any> dumpAll(data: List<T>): String {
-		TODO("not implemented")
+		throw UnsupportedOperationException("Could not find suitable methods to delegate in yamlMapper.")
 	}
 	
 	override fun <T : Any> dumpAll(data: List<T>, file: File) {
-		TODO("not implemented")
+		throw UnsupportedOperationException("Could not find suitable methods to delegate in yamlMapper.")
 	}
 	
 	override fun <T : Any> dumpAll(data: List<T>, writer: Writer) {
-		TODO("not implemented")
+		throw UnsupportedOperationException("Could not find suitable methods to delegate in yamlMapper.")
 	}
 }