+++ learn-kotlin/src/test/kotlin/com/windea/study/kotlin/HelloWorldTest.kt
@@ -1,0 +1,10 @@
+package com.windea.study.kotlin
+
+import org.junit.*
+
+class HelloWorldTest {
+	@Test
+	fun helloWorld() {
+		println("Hello world!")
+	}
+}